Add Kernel#squawk extension
Squawk acts as a guard clause for argument type assertions.  Default behavior is to throw a Squawk::ArgumentTypeMismatch error.
To assert that an argument matches a given type pass the subject (argument) and a type...

```ruby
class Person
  def fullname segments
    squawk segments, Array
  end
end
```

Optionally, pass a block to run more complex assertions.  The expectation is that the block will return `false` if the subject (argument) is invalid...

```ruby
class Person
  def fullname
    squawk segments do |sub|
      sub.first == 'John'
    end
  end
end
```